+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [unreleased]
+- Added cache writing to disk
+- Added `config` method
+
 ## [2.0.0] - 2017-08-13
 - Changed `set` method: last argument now is an object, see documentation.
 - Added expiry date